+++ example/lib/constants.dart
@@ -30,4 +30,6 @@ class Constants {
   static String logo = "logo.png";
   static String placeholderUrl =
       "https://imgix.bustle.com/uploads/image/2020/8/5/23905b9c-6b8c-47fa-bc0f-434de1d7e9bf-avengers-5.jpg";
+  static String elephantDreamStreamUrl =
+      "http://cdn.theoplayer.com/video/elephants-dream/playlist.m3u8";
 }